+++ SimpleCardBrand.exs
@@ -47,6 +47,11 @@ defmodule SimpleCardBrand do
     _card_brand(String.codepoints(pan_head), pan_length)
   end
 
+  # China T-Union
+  defp _card_brand(["3", "1" | _], 19)  do
+    {:ok, :chinatunion}
+  end
+
   # American Express
   defp _card_brand([ "3", second | _ ], 15) when second in ["4", "7"] do
     {:ok, :americanexpress}
@@ -73,6 +78,12 @@ defmodule SimpleCardBrand do
   defp _card_brand([ "4" | _ ], pan_length) when pan_length in [13, 16, 19] do
     {:ok, :visa}
   end
+
+  # China UnionPay: Before Discover 622
+  defp _card_brand(["6", "2" | _], pan_length) when pan_range(pan_length, 16, 19) do
+    {:ok, :chinaunionpay}
+  end
+
 
   # Discover
   defp  _card_brand(["6", "0", "1", "1" | _], pan_length) when pan_range(pan_length, 16, 19) do
@@ -106,7 +117,7 @@ defmodule SimpleCardBrand do
 end
 
 SimpleCardBrand.card_brand("2221000", 16)  # Mastercard
-SimpleCardBrand.card_brand("2220000", 16)  # ErlangError
+SimpleCardBrand.card_brand("2220000", 16)  # Error
 SimpleCardBrand.card_brand("4111111111111111", 16)  # Visa
 SimpleCardBrand.card_brand("3456", 15)  # American Express
 SimpleCardBrand.card_brand("3567", 15)  # Error
@@ -114,3 +125,5 @@ SimpleCardBrand.card_brand("648123", 17) # Discover
 SimpleCardBrand.card_brand("62201111", 17) # Error
 SimpleCardBrand.card_brand("622126", 17) # Discover
 SimpleCardBrand.card_brand("4111111111111111")  # Visa
+SimpleCardBrand.card_brand("622345", 17)  # China UnionPay
+SimpleCardBrand.card_brand("31456", 19)  # China T-Union