+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.0.3 - WIP
 - Make `pan_length` first parameter of `pan_range`.
+- China UnionPay, China T-Union
 
 ## v0.0.2 - 2023-08-03
 - Initial release.